Move openrouter provider
