+++ aisdk/ai/provider/internal/openrouter/client/prompt.go
