+++ aisdk/ai/provider/internal/openrouter/client/prompt_json.go
