+++ aisdk/ai/provider/internal/openrouter/client/settings.go
