+++ aisdk/ai/provider/internal/openrouter/client/settings_test.go
