+++ aisdk/ai/provider/internal/openrouter/codec/decode_finish_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"go.jetify.com/ai/api"
-	"go.jetify.com/ai/provider/openrouter/internal/client"
+	"go.jetify.com/ai/provider/internal/openrouter/client"
 )
 
 func TestDecodeFinishReason(t *testing.T) {
