+++ aisdk/ai/provider/internal/openrouter/convert_to_openrouter_completion_prompt.go
