+++ aisdk/ai/provider/internal/openrouter/model/doc.go
