+++ aisdk/ai/provider/internal/openrouter/ptr.go
